move build-sw to end
