+++ package.json
@@ -2,7 +2,7 @@
   "name": "web.dev",
   "version": "2.0.0",
   "scripts": {
-    "build": "npm-run-all clean sass gulp rollup eleventy",
+    "build": "npm-run-all clean sass gulp rollup eleventy rollup-sw",
     "clean": "rm -rf dist",
     "coverage:integration": "c8 npm-run-all test:lib test:unit test:integration",
     "coverage:unit": "c8 npm run test:unit",
@@ -13,7 +13,7 @@
     "debug:test:lib": "TEST_MODE=debug npm-run-all clean rollup --parallel watch:rollup karma",
     "debug:test:unit": "node --inspect-brk ./node_modules/.bin/mocha ./test/unit",
     "deploy": "ELEVENTY_ENV=prod npm run build && node index-algolia.js && gcloud app deploy --project web-dev-production-1 --quiet",
-    "dev": "ELEVENTY_ENV=dev npm-run-all clean rollup sass --parallel --race watch:rollup watch:sass watch:eleventy watch:gulp start",
+    "dev": "ELEVENTY_ENV=dev npm-run-all clean rollup rollup-sw sass --parallel --race watch:rollup watch:sass watch:eleventy watch:gulp watch:rollup-sw start",
     "eleventy": "eleventy --quiet",
     "gulp": "npx gulp build",
     "karma": "karma start",
@@ -24,6 +24,7 @@
     "percy": "PERCY=true npm-run-all build --parallel --race start snapshots",
     "precommit": "node ./tools/update-updated/index.js",
     "rollup": "node build.js",
+    "rollup-sw": "node build-sw.js",
     "sass": "node ./compile-css.js",
     "snapshots": "percy exec --config ./tools/percy/.percy.conf.yml -- node tools/percy/snapshots.js",
     "stage:personal": "ELEVENTY_ENV=dev npm run build && gcloud app deploy --project web-dev-staging --quiet --no-promote",
@@ -36,6 +37,7 @@
     "watch:eleventy": "eleventy --watch --quiet",
     "watch:gulp": "npx gulp build && gulp watch",
     "watch:rollup": "chokidar \"src/lib/**/*.{js,scss}\" -c \"npm run rollup\"",
+    "watch:rollup-sw": "chokidar \"src/lib/**/*.js\" -c \"npm run rollup-sw\"",
     "watch:sass": "chokidar \"src/styles/**/*.scss\" -c \"npm run sass\"",
     "watch:test:lib": "TEST_MODE=dev npm-run-all clean rollup --parallel watch:rollup karma"
   },
